removing more old code + starting to move over dialog to Vue.js, #20

+++ src/main/kotlin/org/libraryweasel/music/abc/ABCXodusStore.kt
@@ -68,7 +68,7 @@ class ABCXodusStore : ABCManager {
     override fun removeABCDocument(user: User, id: Long): Boolean {
         return entityStore.instance.computeInTransaction { txn ->
             val result = fetchQueryEntity(txn, id)
-            if (result != null) {
+            if (result != null && result.getProperty("username") as String == username(user)) {
                 result.delete()
                 true
             } else {
